+++ addOns/ascanrules/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Improved the following scan rules by using time-based linear regression tests:
   - SQL Injection - Hypersonic SQL
   - SQL Injection - MsSQL
+  - Server Side Template Injection (Blind)
 
 ## [59] - 2023-12-07
 ### Added